PDF_Intergration

+++ static/tools/mpt.py
@@ -620,16 +620,18 @@ class MPT(Debuggable):
                 shutil.rmtree(os.path.join(p.get('path'), uid))
         else:
             t_path.append(prefix + '.' + out_type)
-            #print t_path
             p_path = self.gv.create_dirs_recursive(project_path)
             temp_path = os.path.dirname(os.path.abspath(SEP.join(t_path)))
             for filename in os.listdir(temp_path):
-                f_path = '{}{}{}'.format(p_path, SEP, filename)
-                if os.path.exists(SEP.join(t_path)):
-                    os.rename(SEP.join(t_path), f_path)
+                #f_path = '{}{}{}'.format(p_path, SEP, filename)
+                f_path = '{}{}{}.{}'.format(p_path, SEP, prefix, out_type)
+                os.rename(SEP.join(t_path), f_path)
+                if not p['typesetters'][p_id].get('merge'):
+                    shutil.rmtree(os.path.join(p.get('path'), uid))
 
 
-            shutil.rmtree(os.path.join(p.get('path'), uid))
+
+        #shutil.rmtree(os.path.join(p.get('path'), uid))
 
         #self.debug.print_console(self, '{}  {}'.format(self.gv.OUTPUT,f_path))
